requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
openai-whisper==20231117
torch==2.2.2
groq==0.9.0
yt-dlp==2024.5.27
httpx==0.27.0
